+++ src/read-more.js
@@ -2,7 +2,7 @@
  * ReadMore.js
  * JavaScript library that adds a 'read more' functionality on the text blocks that is applied to.
  *
- * @version 1.2.0
+ * @version 2.0.0
  * @author George Raptis <georapbox@gmail.com>
  * @repository https://github.com/georapbox/ReadMore.js.git
  * @license MIT
@@ -15,7 +15,7 @@
   } else {
     context[name] = definition(name, context);
   }
-}('$readMoreJS', this, function () {
+}('$ReadMoreJS', this, function () {
   'use strict';
 
   function extend() {
@@ -39,18 +39,17 @@
   }
 
   function generateTrimmed(str, wordsNum) {
-    return str.split(/\s+/).slice(0, wordsNum).join(' ') + '...';
+    return str.split(/\s+/).slice(0, wordsNum).join(' ') + '... ';
   }
 
   function init(options) {
     var defaults = {
       target: '',
-      numOfWords: 50,
+      wordsCount: 50,
       toggle: true,
-      moreLink: 'read more...',
+      moreLink: 'read more',
       lessLink: 'read less',
-      linkClass: 'rm-link',
-      containerClass: ''
+      linkClass: 'rm-link'
     };
 
     options = extend({}, defaults, options);
@@ -58,52 +57,46 @@
     var targets = document.querySelectorAll(options.target);
     var initArr = [];
     var trimmedArr = [];
-    var targetContent, trimmedTargetContent, targetContentWords, i, j, moreContainer, rmLinks;
+    var targetContent, trimmedTargetContent, targetContentWords, i, j, rmLinks;
 
     function onMoreAnchorClicked(evt) {
       evt.preventDefault();
 
-      var moreLinkID = this.getAttribute('id');
-      var index = moreLinkID.split('_')[1];
+      var linkEl = evt.currentTarget;
+      var linkId = linkEl.getAttribute('id');
+      var index = linkId.split('_')[1];
 
-      if (this.getAttribute('data-clicked') !== 'true') {
+      if (linkEl.getAttribute('data-clicked') !== 'true') {
         targets[index].innerHTML = initArr[index];
 
         if (options.toggle) {
-          this.innerHTML = options.lessLink;
-          this.setAttribute('data-clicked', true);
+          linkEl.innerHTML = options.lessLink;
+          linkEl.setAttribute('data-clicked', true);
+          targets[index].appendChild(linkEl);
         } else {
-          this.innerHTML = '';
+          linkEl.removeEventListener('click', onMoreAnchorClicked);
         }
       } else {
         targets[index].innerHTML = trimmedArr[index];
-        this.innerHTML = options.moreLink;
-        this.setAttribute('data-clicked', false);
+        targets[index].appendChild(linkEl);
+        linkEl.innerHTML = options.moreLink;
+        linkEl.setAttribute('data-clicked', false);
       }
     }
 
     for (i = 0; i < targets.length; i++) {
       targetContent = targets[i].innerHTML;
-      trimmedTargetContent = generateTrimmed(targetContent, options.numOfWords);
+      trimmedTargetContent = generateTrimmed(targetContent, options.wordsCount);
       targetContentWords = countWords(targetContent);
       initArr.push(targetContent);
       trimmedArr.push(trimmedTargetContent);
 
       // Procceed only if the number of words specified by the user is smaller than the number of words the target element has.
-      if (options.numOfWords < targetContentWords - 1) {
-        targets[i].innerHTML = trimmedArr[i];
-        moreContainer = document.createElement('div');
-
-        if (options.containerClass) {
-          moreContainer.className = options.containerClass;
-        }
-
-        moreContainer.innerHTML = '<a href="#" data-readmore="anchor" id="rm-more_' + i + '"'
+      if (options.wordsCount < targetContentWords - 1) {
+        targets[i].innerHTML = trimmedArr[i] + '<a href="#" data-readmore="anchor" data-clicked="false" id="rm-more_' + i + '"'
           + ' class="' + options.linkClass + '">'
           + options.moreLink
           + '</a>';
-
-        targets[i].parentNode.insertBefore(moreContainer, targets[i].nextSibling);
       }
     }
 
@@ -114,19 +107,15 @@
     }
 
     return function destroy() {
-      for (i = 0; i < targets.length; i++) {
-        targets[i].innerHTML = initArr[i];
-      }
-
       for (j = 0; j < rmLinks.length; j++) {
         rmLinks[j].removeEventListener('click', onMoreAnchorClicked);
-        moreContainer = rmLinks[j].parentNode;
-        moreContainer.parentNode && moreContainer.parentNode.removeChild(moreContainer);
+      }
+
+      for (i = 0; i < targets.length; i++) {
+        targets[i].innerHTML = initArr[i];
       }
     };
   }
 
-  return {
-    init: init
-  };
+  return init;
 }));