+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## v2.0.0 - 2021-08-20
+
+### Breaking changes
+- Rename `numOfWords` option to `wordsCount`.
+- Remove the container element of the "read more/less" link along with `containerClass` option and place the link inline by default. You can still use a CSS class to make the link element block using the `linkClass` option.
+- Update the way to initialise the library.  
+  Prior to v2.x: `$readMoreJS.init(options);`.  
+  From v2.x onwards: `$readMoreJS(options);`
+
 ## v1.2.0 - 2021-08-18
 - Export library in UMD format.
 - Provide a way to destroy if no longer needed.